[Refactor] Defining the `Package` interface
This commit refactors the code to add a new `Package` interface
that will be the standard interface for package model-based
landpattern design. The idea is that the user creates a
Package model of a IC, eg a SOIC. Then the user implements the
interface of `Package` consisting of `make-landpattern`
to generate a landpattern architecture.

The generator makes these pieces more composable rather than
creating `pcb-landpattern` definitions directly. There is also
a helper function for this but not the core interface.