[Refactor] Updated the thermal pad generator
I've simplified this function so that it just accepts a
pad-num instead of the total pins in the package. The user can decide
what pad number to give it and it could be `total + 1`

Additionally, I've added an optional arg for the pose placement with
default at 0,0 as it is currently.