Added Virtual LandPattern Model
This commit introduces a virtual landpattern tree model that allows
for the composition of multiple landpatterns. This model
allows for better introspection, separation, and composability
of different features for the landpatterns.

This concept is primarily a work around for the fact that writing
definitions to ESIR is an immutable process. This tree structure
will allow us to delay writing the ESIR definitions until the last
step of generating a landpattern.