[Refactor] Moved `density-level` as a defmulti of `Package`
This simplifies the code and removes the need to pass the
`density-level:DensityLevel` parameter around to all
methods and generators.

This isn't optimal - my original goal was that this density level
wasn't defined in the `Package` because it is part of the landpattern
and not part of the physical features of the package.

However, this is a practical decision to reduce the amount of typing
and work needed to use the interfaces.