[Refactor] Simplified the SymbolDefn Interface
This removes the `build-designator` method in favor
of the singular `build-params` method. I found that I
was not using the `build-params` because the functions to
compute the bounds of the glyphs either:

1.  Didn't handle the `Text` shape (I've now fixed this).
2.  The `Text` shapes were too big or threw off the bounds calculations.

By making all of the designator or value parameters get
generated in one function, I can avoid this and simplify the interface to these types.