Added DIAC and TRIAC symbols
I've tried a different structure for this implementation
where I'm using the `SymbolNode` to build up the geometry one
glyph at a time. This prevents the need for explicit `Vector`
and other coding. It doesn't prevent the need to create shapes and
add them explicitly with `add-glyph` but seems like it might be a
better way than the previous functions that just generate shapes.