Added initial implementation of the Microstrip transmission line.
This provides the calculation of the effective relative permittivity
from Schneider with the microstrip impedance equation from Wheeler.

This calculation is from:

Wadell, Brian C. "Transmission Line Design Handbook"
Section 3.5.1

Note: I've only implemented the DC version of this so far - we will
need to implement the dispersion relationship variant in the future.