[WIP] Initial Interface for Box Symbol v2
This is intended to be a pointer in the right direction. There are
probably things that are wrong, there may be conceptual issues with
working with `Instantiable|Self`.

My goal is to avoid `inside pcb-component` generator like functions
as much as possible.

We will likely need some method of accessing the `pin-properties`
table from the `Instantiable|Self` parameter or something.