[Refactor] Consolidated the landpattern and symbol frameworks
This refactors the `jsl/landpatterns` and `jsl/symbols` packages to
be the primary interface for users of this library.

I've created two files in each subdir, `framework.stanza` and `generators.stanza`

The Framework package imports the pkgs needed when building new generators.
The Generators package imports the pkgs needed when creating new components.

The top-level package forwards both of these.