Added column omitition for the grid numbering
In some packages, particularlly BGAs composed of pad matrices with
differing pitches, the grid may need to skip some column numbers.
This is a similar but simpler problem to the omitted row letters.

This adds the ability for the user to pass a column omit list
so that the grid numbering will skip some columns when necessary.