[Refactor] Modified shape/hole generator behavior to simplify
Previously, the code was checking `active?` in the shape and
hole generator functions which was resulting in a lot of
copying and pasting.

I'm refactoring this so that we check `active?` in the
`pad-generator` functions so that this gets consolidated.

This has a nice effect of reducing the complexity of the shape
and hole generator functions.