Added more introspection of components, modules, and ports
This adds features for accessing components and pads from ports.
I've refined the existing `get-pads-from-port` so that we don't
need to pass an argument anymore.

I've achieved this by creating a `get-component-from-port` which
can get the component instance associated with a port. We can
then create the pad mapping from that component.

I've added some unit tests for the functionality to confirm that
works as expected.