Updating the silkscreen to use an interface type
This is baby step in a direction that will allow the user
to customize the silkscreen for their land patterns.

This provides a base interface type for silkscreen outlines and
then provides several implementations for various different types of
ICs.

In future iterations the user will be able to pass custom versions in.