[Refactor] Removed the `to-lp-pd-seq` function
I had previously added this because the `bounds` function
was looking for a `LandPatternPad` type as its argument.
This is yet another case where attempting to use types is
difficult if not impossible in JITX. I've converted `bounds`
to use `JITXObject` instead and this removes the
need for the seq function.