Updated the MDI 1000Base-T Supports Generators
When attempting to create the supports for the KSZ9563
ethernet ports - I found that I needed more flexible
control. I needed to construct a composite bundle
that combined the MDI and the LED control pins.

By passing in a bundle argument, I can create this
composite but still implement the MDI twisted
pair mappings.